Create an example of composition.
Composition occurs when two classes are related but there is no
inheritance. The class Repository contains a collection of packages.
Each package is an individual object.

This type of a relation is a "has a". Another example of composition is
a Car is a Vehicle, but a Car has an Engine. Car and Engine are related
but there is no inheritance.

For inheritance, it is a "is a " relation.